Rename /apis to /api
and make relevant necessary changes

Partially-Closes: #1329
