+++ controllers/servicediscovery/servicediscovery_controller_test.go
@@ -23,7 +23,7 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	operatorv1 "github.com/openshift/api/operator/v1"
-	submariner_v1 "github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
+	submariner_v1 "github.com/submariner-io/submariner-operator/api/submariner/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
