+++ go.mod
@@ -38,7 +38,8 @@ require (
 	github.com/submariner-io/lighthouse v0.12.0-m0
 	github.com/submariner-io/shipyard v0.12.0-m0
 	github.com/submariner-io/submariner v0.12.0-m0
-	github.com/submariner-io/submariner-operator/apis v0.0.0-20210817145008-861856b068a1
+	github.com/submariner-io/submariner-operator/api v0.0.0-20210817145008-861856b068a1
+	github.com/submariner-io/submariner-operator/apis v0.0.0-20211108185031-39851b39dec6
 	github.com/submariner-io/submariner/pkg/apis v0.0.0-20210816153739-c8e6654e3930
 	github.com/ulikunitz/xz v0.5.10 // indirect
 	github.com/uw-labs/lichen v0.1.4
@@ -78,4 +79,4 @@ replace (
 )
 
 // Local packages
-replace github.com/submariner-io/submariner-operator/apis => ./apis
+replace github.com/submariner-io/submariner-operator/api => ./api
