+++ pkg/client/clientset/versioned/typed/submariner/v1alpha1/broker.go
@@ -22,7 +22,7 @@ import (
 	"context"
 	"time"
 
-	v1alpha1 "github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
+	v1alpha1 "github.com/submariner-io/submariner-operator/api/submariner/v1alpha1"
 	scheme "github.com/submariner-io/submariner-operator/pkg/client/clientset/versioned/scheme"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	types "k8s.io/apimachinery/pkg/types"
