+++ pkg/client/clientset/versioned/typed/submariner/v1alpha1/fake/fake_broker.go
@@ -21,7 +21,7 @@ package fake
 import (
 	"context"
 
-	v1alpha1 "github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
+	v1alpha1 "github.com/submariner-io/submariner-operator/api/submariner/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	labels "k8s.io/apimachinery/pkg/labels"
 	schema "k8s.io/apimachinery/pkg/runtime/schema"
