+++ pkg/client/clientset/versioned/typed/submariner/v1alpha1/submariner_client.go
@@ -19,7 +19,7 @@ limitations under the License.
 package v1alpha1
 
 import (
-	v1alpha1 "github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
+	v1alpha1 "github.com/submariner-io/submariner-operator/api/submariner/v1alpha1"
 	"github.com/submariner-io/submariner-operator/pkg/client/clientset/versioned/scheme"
 	rest "k8s.io/client-go/rest"
 )
