+++ pkg/images/images.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 	"strings"
 
-	apis "github.com/submariner-io/submariner-operator/apis/submariner/v1alpha1"
+	apis "github.com/submariner-io/submariner-operator/api/submariner/v1alpha1"
 	"github.com/submariner-io/submariner-operator/pkg/names"
 
 	v1 "k8s.io/api/core/v1"
